# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Позволяет управлять сегментами и пользователями в них.",
    "title": "Сервис сегментов пользователя",
    "version": "0.0.1"
  },
  "host": "localhost:8080",
  "paths": {
    "/segments": {
      "post": {
        "description": "Создает новый сегмент",
        "tags": [
          "segments"
        ],
        "operationId": "createSegment",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateSegmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createSegmentResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/segments/{slug}": {
      "delete": {
        "description": "Удаляет сегмент",
        "tags": [
          "segments"
        ],
        "operationId": "deleteSegment",
        "parameters": [
          {
            "type": "string",
            "example": "avito_discount_70",
            "x-go-name": "Slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/deleteSegmentResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/users/{userID}/segments": {
      "get": {
        "description": "Возвращает сегменты, в которых состоит пользователь",
        "tags": [
          "user-segments"
        ],
        "operationId": "getUserSegments",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "example": 89,
            "x-go-name": "UserID",
            "name": "userID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getUserSegmentsResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "patch": {
        "description": "Обновляет сегменты у пользователя",
        "tags": [
          "user-segments"
        ],
        "operationId": "updateUserSegments",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserSegmentsRequest"
            }
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "UserID",
            "name": "userID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateUserSegmentsResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateSegmentRequest": {
      "type": "object",
      "properties": {
        "slug": {
          "type": "string",
          "x-go-name": "Slug",
          "example": "avito_discount_90"
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "CreateSegmentResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "x-go-name": "Ok"
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "DeleteSegmentResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "x-go-name": "Ok"
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "GetUserSegmentsResponse": {
      "type": "object",
      "properties": {
        "slugs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Slugs",
          "example": [
            "avito_discount_70",
            "avito_discount_50"
          ]
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "UpdateUserSegmentsRequest": {
      "type": "object",
      "properties": {
        "add_slug": {
          "description": "добавляет пользователя в эти сегменты",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "AddSlug",
          "example": [
            "avito_discount_70",
            "avito_discount_50"
          ]
        },
        "delete_slug": {
          "description": "удаляет пользователя из этих сегментов",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "DeleteSlug",
          "example": [
            "avito_discount_30"
          ]
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    },
    "UpdateUserSegmentsResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "x-go-name": "Ok"
        }
      },
      "x-go-package": "github.com/lyuvolkova/avitoTest-dynSeg/internal/pkg/server"
    }
  },
  "responses": {
    "createSegmentResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/CreateSegmentResponse"
      }
    },
    "deleteSegmentResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/DeleteSegmentResponse"
      }
    },
    "errorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
    "getUserSegmentsResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GetUserSegmentsResponse"
      }
    },
    "updateUserSegmentsResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UpdateUserSegmentsResponse"
      }
    }
  }
}